Fix issue where row update is processed before value is persisted
Occasionally it seems MUI Grid's `processRowUpdate` is called before the async `setEditCellValue` resolves. This means the `newRow` parameter of `processRowUpdate` is stale and contains the old value. Adding debounce seems to fix the issue but it's hard to test due to the sporadic nature of the error. I expect it's some kind of race condition in the MUI code.